requirements.txt
spacy
transformers
torch
google-generativeai
fastapi
uvicorn
python-dotenv
